requirements.txt
Django
psycopg2-binary
djangorestframework
django-cors-headers
gunicorn
python-decouple
django-environ
EOL
python-dotenv
djangorestframework-stubs
djangorestframework-simplejwt
twilio
pandas
openpyxl




# python manage.py import_students "Data.xlsx"
